тз10

+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -20,5 +20,5 @@ public class User {
     @PastOrPresent(message = "Дата рождения не может быть в будущем")
     private LocalDate birthday;
 
-    private Set<Integer> friendsId;
+    private Set<Long> friendsId;
 }
